AEROGEAR-8188 - audit logging module improvements
