+++ packages/apollo-voyager-tools/package.json
@@ -9,6 +9,9 @@
     "test": "ava '*.test.ts' '**/*.test.ts'",
     "compile:clean": "tsc --build tsconfig.json --clean"
   },
+  "dependencies": {
+    "graphql-tools": "^4.0.3"
+  },
   "devDependencies": {
     "ava": "1.0.0-rc.2",
     "ts-node": "^7.0.1",
