+++ packages/apollo-voyager-tools/src/index.ts
@@ -1,1 +1,2 @@
 export * from './buildPath'
+export * from './wrapResolvers'
